[Fix] Made elements in Dashboard.jsx into seperate controller & views
- Created statCards component
- Created dashboardGraphes component
- Updated Dashboard.jsx to make it work with both components
- Fixed Backend Admin getters/manipulators
