+++ services/backend/API/admin/reporting/util/growthRateByMonth.js
@@ -1,11 +1,10 @@
 const moment = require("moment");
-const Sequelize = require("sequelize");
 const Models = require("../../../../models");
 const Profiles = Models.profile; // Profiles Table
 
 const growthRateByMonth = async () => {
   // Object Structure: {Year, Data: [{Month, # of Occurrences}]}
-  let growthRate = [];
+  let monthlyGrowthRate = [];
 
   const profiles = await Profiles.findAll({
     attributes: ["id", "createdAt"]
@@ -18,26 +17,26 @@ const growthRateByMonth = async () => {
 
     const profileYear = profileCreatedAt.year();
 
-    const indexYear = growthRate.findIndex(
+    const indexYear = monthlyGrowthRate.findIndex(
       object => object.year == profileYear
     );
 
     if (indexYear != -1) {
-      const indexMonth = growthRate[indexYear].data.findIndex(
+      const indexMonth = monthlyGrowthRate[indexYear].data.findIndex(
         object => object.month == profileMonth
       );
 
       if (indexMonth != -1) {
-        growthRate[indexYear].data[indexMonth].count =
-          growthRate[indexYear].data[indexMonth].count + 1;
+        monthlyGrowthRate[indexYear].data[indexMonth].count =
+          monthlyGrowthRate[indexYear].data[indexMonth].count + 1;
       } else {
-        growthRate[indexYear].data.push({
+        monthlyGrowthRate[indexYear].data.push({
           month: profileMonth,
           count: 1
         });
       }
     } else {
-      growthRate.push({
+      monthlyGrowthRate.push({
         year: profileYear,
         data: [
           {
@@ -49,11 +48,11 @@ const growthRateByMonth = async () => {
     }
   });
 
-  for (let i = 0; i < growthRate.length; i++) {
-    growthRate[i].data.sort((a, b) => a.month - b.month);
+  for (let i = 0; i < monthlyGrowthRate.length; i++) {
+    monthlyGrowthRate[i].data.sort((a, b) => a.month - b.month);
   }
 
-  growthRate = growthRate.map((entry, index) => {
+  monthlyGrowthRate = monthlyGrowthRate.map(entry => {
     let newData = [];
     let newMonth = [];
 
@@ -71,7 +70,76 @@ const growthRateByMonth = async () => {
     return entry;
   });
 
-  return growthRate;
+  const months = moment.monthsShort();
+
+  monthlyGrowthRate = monthlyGrowthRate.map(entry => {
+    let addData = [];
+
+    for (let i = 0; i < 12; i++) {
+      addData.push({
+        month: i,
+        monthName: months[i],
+        count: entry.data[i].count
+      });
+    }
+    entry.data = addData;
+    return entry;
+  });
+
+  // Growth Rate By Month Graph Data:
+  let graphicalData = [];
+
+  monthlyGrowthRate = monthlyGrowthRate.map(entry => {
+    for (let i = 0; i < 12; i++) {
+      graphicalData.push({
+        year: entry.year.toString(),
+        monthNumber: i,
+        monthName: entry.data[i].monthName,
+        count: entry.data[i].count
+      });
+    }
+    return entry;
+  });
+
+  // Users Added This Month & Growth Rate Percentage:
+  const current_year = moment().year();
+
+  const current_month = moment().month();
+
+  const indexYear = monthlyGrowthRate.findIndex(
+    object => object.year === current_year
+  );
+
+  const current_month_additions = monthlyGrowthRate[indexYear].data.find(
+    object => object.month === current_month
+  );
+
+  let previous_month_additions = {};
+
+  if (current_month === 0) {
+    previous_month_additions = monthlyGrowthRate[indexYear - 1].data.find(
+      object => object.month === 11
+    );
+  } else {
+    previous_month_additions = monthlyGrowthRate[indexYear].data.find(
+      object => object.month === current_month - 1
+    );
+  }
+  let growthRateFromPreviousMonth = Math.round(
+    ((current_month_additions.count - previous_month_additions.count) /
+      previous_month_additions.count) *
+      100
+  );
+
+  if (growthRateFromPreviousMonth === Infinity) {
+    growthRateFromPreviousMonth = "N/A";
+  }
+
+  return {
+    graphicalData,
+    current_month_additions,
+    growthRateFromPreviousMonth
+  };
 };
 
 module.exports = growthRateByMonth;
