+++ services/frontend-v3/src/config.js
@@ -3,5 +3,5 @@ export default {
   backendAddressDev:
     "http://upskill-backend-upskill.apps.dev.openshift.ised-isde.canada.ca/",
   backendAddress: "http://localhost:8080/"
-  // "http://upskill-backend-uat-upskill.apps.dev.openshift.ised-isde.canada.ca/"
+  //"http://upskill-backend-uat-upskill.apps.dev.openshift.ised-isde.canada.ca/"
 };
