+++ services/frontend-v3/src/pages/admin/Dashboard.jsx
@@ -1,19 +1,9 @@
 import React from "react";
 import AppLayout from "../../components/layouts/appLayout/AppLayout";
+import StatCards from "../../components/statCards/StatCards";
+import DashboardGraphes from "../../components/dashboardGraphes/DashboardGraphes";
 import axios from "axios";
-
-import {
-  EyeInvisibleFilled,
-  RiseOutlined,
-  SolutionOutlined,
-  TeamOutlined,
-  UsergroupAddOutlined,
-  UserOutlined,
-} from '@ant-design/icons';
-
-import { Typography, Skeleton, Statistic, Card, Row, Col } from "antd";
-import moment from "moment";
-import { Chart, Geom, Axis, Tooltip, Coord, Legend } from "bizcharts";
+import { Typography, Skeleton } from "antd";
 import { injectIntl } from "react-intl";
 import config from "../../config";
 
@@ -43,6 +33,8 @@ class AdminDashboard extends React.Component {
 
   render() {
     const { data, loading } = this.state;
+    const locale = this.props.intl.formatMessage({ id: "language.code" });
+
     if (loading) {
       return (
         <AppLayout>
@@ -50,141 +42,6 @@ class AdminDashboard extends React.Component {
         </AppLayout>
       );
     }
-
-    // Skills Count:
-    const skillsCount = data.skillCount;
-
-    const topFiveSkills = [];
-
-    for (let i = 0; i < 5; i++) {
-      const skill = skillsCount[i];
-
-      topFiveSkills.push({
-        name:
-          this.props.intl.formatMessage({ id: "language.code" }) === "en"
-            ? skill.descriptionEn
-            : skill.descriptionFr,
-        count: parseInt(skill.countOccurences)
-      });
-    }
-
-    // Competencies Count:
-    const competenciesCount = data.compCount;
-
-    const topFiveCompetencies = [];
-
-    for (let i = 0; i < 5; i++) {
-      const competency = competenciesCount[i];
-
-      topFiveCompetencies.push({
-        name:
-          this.props.intl.formatMessage({ id: "language.code" }) === "en"
-            ? competency.descriptionEn
-            : competency.descriptionFr,
-        count: parseInt(competency.countOccurences)
-      });
-    }
-
-    // Development Goals Count:
-    const developmentGoalsCount = data.developCount;
-
-    const topFiveDevelopmentGoals = [];
-
-    for (let i = 0; i < 5; i++) {
-      const developmentGoal = developmentGoalsCount[i];
-
-      topFiveDevelopmentGoals.push({
-        name:
-          this.props.intl.formatMessage({ id: "language.code" }) === "en"
-            ? developmentGoal.descriptionEn
-            : developmentGoal.descriptionFr,
-        count: parseInt(developmentGoal.countOccurences)
-      });
-    }
-
-    // Growth Rate By Month:
-    let monthlyGrowthRate = data.growthRateByMonth;
-
-    const months = moment.monthsShort();
-
-    monthlyGrowthRate = monthlyGrowthRate.map((entry, index) => {
-      let addData = [];
-
-      for (let i = 0; i < 12; i++) {
-        addData.push({
-          month: i,
-          monthName: months[i],
-          count: entry.data[i].count
-        });
-      }
-      entry.data = addData;
-      return entry;
-    });
-
-    let graphicalData = [];
-
-    monthlyGrowthRate = monthlyGrowthRate.map((entry, index) => {
-      for (let i = 0; i < 12; i++) {
-        graphicalData.push({
-          year: entry.year.toString(),
-          monthNumber: i,
-          monthName: entry.data[i].monthName,
-          count: entry.data[i].count
-        });
-      }
-      return entry;
-    });
-
-    // Users Added This Month & Growth Rate Percentage:
-    const current_year = moment().year();
-
-    const current_month = moment().month();
-
-    const indexYear = monthlyGrowthRate.findIndex(
-      object => object.year === current_year
-    );
-
-    const current_month_additions = monthlyGrowthRate[indexYear].data.find(
-      object => object.month === current_month
-    );
-
-    let previous_month_additions = {};
-
-    if (current_month === 0) {
-      previous_month_additions = monthlyGrowthRate[indexYear - 1].data.find(
-        object => object.month === 11
-      );
-    } else {
-      previous_month_additions = monthlyGrowthRate[indexYear - 1].data.find(
-        object => object.month === 11
-      );
-    }
-    const growthRate = Math.round(
-      ((current_month_additions.count - previous_month_additions.count) /
-        previous_month_additions.count) *
-        100
-    );
-
-    // Graph Columns:
-    const popularSkillsColumns = {
-      name: { alias: "Skill Name" },
-      count: { alias: "Number of users" }
-    };
-
-    const popularCompetenciesColumns = {
-      name: { alias: "Competency Name" },
-      count: { alias: "Number of users" }
-    };
-
-    const popularDevelopmentGoalsColumns = {
-      name: { alias: "Development Goal Name" },
-      count: { alias: "Number of users" }
-    };
-
-    const growthRateByMonthColumns = {
-      monthName: { range: [0, 1] },
-      count: { alias: "Number of users" }
-    };
 
     return (
       <AppLayout
@@ -194,214 +51,8 @@ class AdminDashboard extends React.Component {
         displaySideBar={true}
       >
         <Title>Admin Dashboard</Title>
-        <Row gutter={[8, 8]} type="flex">
-          <Col span={4}>
-            <Card hoverable style={{ height: "100%" }}>
-              <Statistic
-                title={this.props.intl.formatMessage({
-                  id: "admin.dashboard.total.users",
-                  defaultMessage: "Total Users"
-                })}
-                value={data.dashboardCount.user}
-                valueStyle={{ color: "#3f8600" }}
-                prefix={<TeamOutlined />}
-              />
-            </Card>
-          </Col>
-          <Col span={4}>
-            <Card hoverable style={{ height: "100%" }}>
-              <Statistic
-                title={this.props.intl.formatMessage({
-                  id: "admin.dashboard.inactive.users",
-                  defaultMessage: "Inactive Users"
-                })}
-                value={data.dashboardCount.inactive}
-                valueStyle={{ color: "#515052" }}
-                prefix={<UserOutlined />}
-              />
-            </Card>
-          </Col>
-          <Col span={4}>
-            <Card hoverable style={{ height: "100%" }}>
-              <Statistic
-                title={this.props.intl.formatMessage({
-                  id: "admin.dashboard.flagged.profiles",
-                  defaultMessage: "Hidden Profiles"
-                })}
-                value={data.dashboardCount.flagged}
-                valueStyle={{ color: "#cf1322" }}
-                prefix={<EyeInvisibleFilled />}
-              />
-            </Card>
-          </Col>
-          <Col span={4}>
-            <Card hoverable style={{ height: "100%" }}>
-              <Statistic
-                title={this.props.intl.formatMessage({
-                  id: "admin.dashboard.ex.feeders",
-                  defaultMessage: "Total Ex Feeders"
-                })}
-                value={data.dashboardCount.exFeeder}
-                valueStyle={{ color: "#82A7A6" }}
-                prefix={<SolutionOutlined />}
-              />
-            </Card>
-          </Col>
-          <Col span={4}>
-            <Card hoverable style={{ height: "100%" }}>
-              <Statistic
-                title={
-                  this.props.intl.formatMessage({
-                    id: "admin.dashboard.monthly.added",
-                    defaultMessage: "New Users - " + moment().format("MMMM")
-                  }) + moment().format("MMMM")
-                }
-                value={current_month_additions.count}
-                valueStyle={{ color: "#CD8FD6" }}
-                prefix={<UsergroupAddOutlined />}
-              />
-            </Card>
-          </Col>
-          <Col span={4}>
-            <Card hoverable style={{ height: "100%" }}>
-              <Statistic
-                title={
-                  this.props.intl.formatMessage({
-                    id: "admin.dashboard.growth.rate.percentage",
-                    defaultMessage: "Growth Rate - " + moment().format("MMMM")
-                  }) + moment().format("MMMM")
-                }
-                value={growthRate}
-                valueStyle={{ color: "#FF934F" }}
-                prefix={<RiseOutlined />}
-                suffix="%"
-              />
-            </Card>
-          </Col>
-        </Row>
-        <Row gutter={[8, 8]}>
-          <Col span={8}>
-            <Card
-              hoverable
-              title={this.props.intl.formatMessage({
-                id: "admin.dashboard.popular.skills",
-                defaultMessage: "Popular Skills"
-              })}
-            >
-              <Chart
-                data={topFiveSkills}
-                scale={popularSkillsColumns}
-                padding="auto"
-                forceFit
-              >
-                <Axis name="name" visible={false} />
-                <Axis name="count" title />
-                <Coord scale={[0.7, 0.9]} />
-                <Legend
-                  position="top"
-                  dy={-20}
-                  textStyle={{ fontSize: "12" }}
-                />
-                <Tooltip />
-                <Geom type="interval" position="name*count" color="name" />
-              </Chart>
-            </Card>
-          </Col>
-          <Col span={8}>
-            <Card
-              hoverable
-              title={this.props.intl.formatMessage({
-                id: "admin.dashboard.popular.competencies",
-                defaultMessage: "Popular Competencies"
-              })}
-            >
-              <Chart
-                data={topFiveCompetencies}
-                scale={popularCompetenciesColumns}
-                padding="auto"
-                forceFit
-              >
-                <Axis name="name" visible={false} />
-                <Axis name="count" title />
-                <Coord scale={[0.7, 0.9]} />
-                <Legend
-                  position="top"
-                  dy={-20}
-                  textStyle={{ fontSize: "11" }}
-                />
-                <Tooltip />
-                <Geom type="interval" position="name*count" color="name" />
-              </Chart>
-            </Card>
-          </Col>
-          <Col span={8}>
-            <Card
-              hoverable
-              title={this.props.intl.formatMessage({
-                id: "admin.dashboard.popular.development.goals",
-                defaultMessage: "Popular Development Goals"
-              })}
-            >
-              <Chart
-                data={topFiveDevelopmentGoals}
-                scale={popularDevelopmentGoalsColumns}
-                padding="auto"
-                forceFit
-              >
-                <Axis name="name" visible={false} />
-                <Axis name="count" title />
-                <Coord scale={[0.7, 0.9]} />
-                <Legend
-                  position="top"
-                  dy={-20}
-                  textStyle={{ fontSize: "12" }}
-                />
-                <Tooltip />
-                <Geom type="interval" position="name*count" color="name" />
-              </Chart>
-            </Card>
-          </Col>
-        </Row>
-        <Row gutter={[8, 8]}>
-          <Col span={24}>
-            <Card
-              hoverable
-              title={this.props.intl.formatMessage({
-                id: "admin.dashboard.growth.rate.by.month",
-                defaultMessage: "Growth Rate By Month"
-              })}
-            >
-              <Chart
-                data={graphicalData}
-                scale={growthRateByMonthColumns}
-                padding="auto"
-                forceFit
-              >
-                <Legend />
-                <Axis name="monthName" />
-                <Axis name="count" />
-                <Tooltip crosshairs={{ type: "y" }} />
-                <Geom
-                  type="line"
-                  position="monthName*count"
-                  size={2}
-                  color={"year"}
-                />
-                <Geom
-                  type="point"
-                  position="monthName*count"
-                  size={4}
-                  shape={"circle"}
-                  color={"year"}
-                  style={{
-                    stroke: "#fff",
-                    lineWidth: 1
-                  }}
-                />
-              </Chart>
-            </Card>
-          </Col>
-        </Row>
+        <StatCards data={data} />
+        <DashboardGraphes data={data} locale={locale} />
       </AppLayout>
     );
   }
